+++ src/pages/PlaylistPage.tsx
@@ -1,15 +1,10 @@
-import { useEffect, useState } from 'react';
+import React from 'react';
 
 import { css } from '@emotion/react';
 import { GoKebabHorizontal, GoStar, GoStarFill } from 'react-icons/go';
 import { RiPlayLargeFill, RiAddLargeLine, RiPencilLine } from 'react-icons/ri';
-import { useParams, useNavigate } from 'react-router-dom';
+import { useParams } from 'react-router-dom';
 
-import {
-  getPlaylistWithUser,
-  deletePlaylist,
-  deleteVideoFromPlaylist,
-} from '@/api/endpoints/playlist';
 import defaultProfileImage from '@/assets/images/default-avatar-man.svg';
 import Button from '@/components/common/buttons/Button';
 import IconButton from '@/components/common/buttons/IconButton';
@@ -19,147 +14,34 @@ import Toast from '@/components/common/Toast';
 import NullBox from '@/components/playlistdetail/nullBox';
 import ThumBoxDetail from '@/components/playlistdetail/thumBoxDetail';
 import VideoBoxDetail from '@/components/playlistdetail/vedieoBoxDetail';
+import useBottomSheet from '@/hooks/useBottomSheet';
+import usePlaylistActions from '@/hooks/usePlaylistAction';
+import usePlaylistData from '@/hooks/usePlaylistData';
+import useUserSession from '@/hooks/useUserSession';
+import useVideoActions from '@/hooks/useVideoAction';
 import Header from '@/layouts/layout/Header';
-import { useToastStore } from '@/store/useToastStore';
-import { useToggleStore } from '@/store/useToggleStore';
 import theme from '@/styles/theme';
-import { PlaylistModel } from '@/types/playlist';
-import { UserModel } from '@/types/user';
 
 const PlaylistPage: React.FC = () => {
-  const { playlistId } = useParams<{ playlistId: string }>(); // URL 파라미터에서 playlistId 추출
-  const [playlist, setPlaylist] = useState<PlaylistModel | null>(null);
-  const [user, setUser] = useState<UserModel | null>(null);
-  const isToggled = useToggleStore((state) => state.isToggled);
-  const toggle = useToggleStore((state) => state.toggle);
-  const showToast = useToastStore((state) => state.showToast);
-  const [isLoading, setIsLoading] = useState(true);
-  const [error, setError] = useState<string | null>(null);
-  const [isBottomSheetOpen, setIsBottomSheetOpen] = useState(false);
-  const navigate = useNavigate();
-  const [bottomSheetContentType, setBottomSheetContentType] = useState<
-    'deleteFromPlaylist' | 'deleteVideo'
-  >('deleteFromPlaylist');
-  const [selectedVideo, setSelectedVideo] = useState<{ videoId: string; title: string } | null>(
-    null
+  const { playlistId } = useParams<{ playlistId: string }>();
+  const { playlist, user, isLoading, error } = usePlaylistData(playlistId);
+  const { handleIconButtonClick, handlePlaylistEdit, handleAddPlaylist, handlePlaylistDelete } =
+    usePlaylistActions();
+  const { selectedVideo, handleVideoDelete, onClickVideoKebob } = useVideoActions(
+    playlist,
+    () => {}
   );
-  // 세션 스토리지에서 userSession 문자열을 가져와서 파싱
-  const userSessionStr = sessionStorage.getItem('userSession');
-  if (!userSessionStr) {
-    throw new Error('세션에 유저 ID를 찾을 수 없습니다.');
-  }
-  const userSession = JSON.parse(userSessionStr);
-  const userId = userSession.uid;
+  const {
+    isBottomSheetOpen,
+    bottomSheetContentType,
+    onClickKebob,
+    handleBottomSheetClose,
+    setBottomSheetContentType,
+    setIsBottomSheetOpen,
+  } = useBottomSheet();
+  const { userId, error: userError } = useUserSession();
 
-  useEffect(() => {
-    async function fetchPlaylistWithUser() {
-      if (!playlistId) {
-        setError('Playlist ID is missing');
-        setIsLoading(false);
-        return;
-      }
-
-      try {
-        const result = await getPlaylistWithUser(playlistId);
-
-        if (result) {
-          setPlaylist(result.playlist);
-          setUser(result.user);
-        } else {
-          setError('Playlist not found');
-        }
-      } catch (err) {
-        console.error('Error fetching playlist:', err); // 로그 추가
-        setError('Failed to fetch playlist');
-      } finally {
-        setIsLoading(false);
-      }
-    }
-
-    fetchPlaylistWithUser();
-  }, [playlistId]);
-
-  const handleIconButtonClick = () => {
-    showToast('내 재생목록에 저장되었습니다.');
-    toggle();
-  };
-
-  const handlePlaylistEdit = () => {
-    console.log('플레이리스트 수정페이지로 이동');
-  };
-  const handleAddPlaylist = () => {
-    console.log('플레이리스트 링크 추가하는 모달 팝업');
-  };
-  const onClickKebob = () => {
-    setBottomSheetContentType('deleteFromPlaylist');
-    setIsBottomSheetOpen(true);
-  };
-
-  const handleBottomSheetClose = () => {
-    setIsBottomSheetOpen(false);
-    setSelectedVideo(null);
-  };
-
-  const handlePlaylistDelete = async (playlistId: string) => {
-    try {
-      setIsLoading(true);
-      await deletePlaylist(playlistId);
-      showToast('플레이리스트가 성공적으로 삭제되었습니다.');
-      navigate(-1); // 이전 페이지로 이동
-    } catch (error) {
-      console.error('Error deleting playlist:', error);
-      showToast('플레이리스트 삭제 중 오류가 발생했습니다.');
-    } finally {
-      setIsLoading(false);
-    }
-  };
-
-  const handleVideoDelete = async () => {
-    if (!playlist || !selectedVideo) {
-      console.error('Playlist or selected video is null');
-      return;
-    }
-
-    console.log(
-      `Attempting to delete video ${selectedVideo.videoId} from playlist ${playlist.playlistId}`
-    );
-
-    try {
-      setIsLoading(true);
-      await deleteVideoFromPlaylist(playlist.playlistId, selectedVideo.videoId);
-
-      setPlaylist((prevPlaylist) => {
-        if (!prevPlaylist) {
-          console.error('Previous playlist is null');
-          return null;
-        }
-        const updatedVideos = prevPlaylist.videos.filter(
-          (video) => video.videoId !== selectedVideo.videoId
-        );
-        console.log('Updated videos:', updatedVideos);
-        return {
-          ...prevPlaylist,
-          videos: updatedVideos,
-          videoCount: updatedVideos.length,
-        };
-      });
-
-      showToast('동영상이 성공적으로 삭제되었습니다.');
-    } catch (error) {
-      console.error('Error in handleVideoDelete:', error);
-      showToast('동영상 삭제 중 오류가 발생했습니다.');
-    } finally {
-      setIsLoading(false);
-      setSelectedVideo(null);
-      handleBottomSheetClose();
-    }
-  };
-  const onClickVideoKebob = (video: { videoId: string; title: string }) => {
-    console.log('Selected video for deletion:', video); // 디버깅을 위한 로그 추가
-    setSelectedVideo(video);
-    setBottomSheetContentType('deleteVideo');
-    setIsBottomSheetOpen(true);
-  };
+  const isToggled = false;
 
   if (isLoading) {
     return (
@@ -168,33 +50,30 @@ const PlaylistPage: React.FC = () => {
       </div>
     );
   }
-  if (!playlist || !user) {
+
+  if (error || userError || !playlist || !user || !userId) {
     return (
       <div>
         <Header customStyle={kebabStyle} />
         <NullBox />
+        {error || userError}
       </div>
     );
-  }
-  if (error) {
-    return <div>{error}</div>;
   }
 
   return (
     <div css={containerStyle}>
-      {playlist.userId === userId ? ( // 여기서 user는 로그인한 사용자
+      {playlist.userId === userId ? (
         <Header Icon={GoKebabHorizontal} customStyle={kebabStyle} onIcon={onClickKebob} />
       ) : (
         <Header />
       )}
-      {playlist && (
-        <ThumBoxDetail
-          playlist={playlist}
-          user={user}
-          profileURL={user.profileImg || defaultProfileImage}
-          onClickProfile={() => console.log('프로필 클릭')}
-        />
-      )}
+      <ThumBoxDetail
+        playlist={playlist}
+        user={user}
+        profileURL={user.profileImg || defaultProfileImage}
+        onClickProfile={() => console.log('프로필 클릭')}
+      />
       <div css={buttonBoxStyle}>
         <Button
           styleType='secondary'
@@ -204,7 +83,7 @@ const PlaylistPage: React.FC = () => {
           <RiPlayLargeFill css={iconStyle} />
           Play all
         </Button>
-        {playlist.userId === userId ? ( // 여기서 user는 로그인한 사용자
+        {playlist.userId === userId ? (
           <IconButton Icon={RiPencilLine} onClick={handlePlaylistEdit} />
         ) : (
           <IconButton Icon={isToggled ? GoStarFill : GoStar} onClick={handleIconButtonClick} />
@@ -215,17 +94,21 @@ const PlaylistPage: React.FC = () => {
           <VideoBoxDetail
             key={video.videoId}
             video={video}
-            type={playlist.userId === userId ? 'host' : 'visitor'} // 로그인한 사용자 아이디 비교해서 값이 참이면 host 다르면 visitor
+            type={playlist.userId === userId ? 'host' : 'visitor'}
             channelName={playlist.userName}
             uploadDate={new Date(playlist.createdAt).toLocaleDateString()}
             onClick={() => console.log(`비디오 클릭됨: ${video.videoId}`)}
-            onClickKebob={() => onClickVideoKebob({ videoId: video.videoId, title: video.title })}
+            onClickKebob={() => {
+              onClickVideoKebob({ videoId: video.videoId, title: video.title });
+              setBottomSheetContentType('deleteVideo');
+              setIsBottomSheetOpen(true);
+            }}
           />
         ))
       ) : (
         <NullBox />
       )}
-      {playlist.userId === userId ? (
+      {playlist.userId === userId && (
         <div css={addButtonContainerStyle}>
           <IconButton
             Icon={RiAddLargeLine}
@@ -233,23 +116,7 @@ const PlaylistPage: React.FC = () => {
             onClick={handleAddPlaylist}
           />
         </div>
-      ) : null}
-      <Toast />
-      {/* <BottomSheet
-        contentType='deleteFromPlaylist'
-        isOpen={isBottomSheetOpen}
-        onClose={handleBottomSheetClose}
-        playlists={[
-          {
-            id: playlist.playlistId,
-            title: playlist.title,
-            isPublic: playlist.isPublic,
-            isBookmarked: false,
-            thumURL: playlist.thumbnailUrl,
-          },
-        ]}
-        onPlaylistClick={handlePlaylistDelete}
-      /> */}
+      )}
       <BottomSheet
         contentType={bottomSheetContentType}
         isOpen={isBottomSheetOpen}
@@ -271,15 +138,19 @@ const PlaylistPage: React.FC = () => {
         onPlaylistClick={handlePlaylistDelete}
         onVideoDelete={handleVideoDelete}
       />
+      <Toast />
     </div>
   );
 };
+
 const containerStyle = css`
   padding-bottom: 90px;
 `;
+
 const kebabStyle = css`
   transform: rotate(90deg);
 `;
+
 const buttonBoxStyle = css`
   margin: 1rem;
   display: flex;
@@ -314,6 +185,7 @@ const addButtonContainerStyle = css`
   height: 1px;
   transform: translateX(-50%);
 `;
+
 const floatAddButtonStyle = css`
   position: absolute;
   right: 1.5rem;
@@ -324,4 +196,5 @@ const floatAddButtonStyle = css`
     transform: translateY(-2px);
   }
 `;
+
 export default PlaylistPage;
